backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

pandas==2.2.1
numpy==1.26.4

reportlab==4.0.7

python-dotenv==1.0.1
gunicorn==21.2.0
